+++ tests/test_burst.py
@@ -137,6 +137,21 @@ def test_validate_bursts_num_swath_pol():
             ['S1_000000_IW1_20200201T000000_VH_0000-BURST'],
         )
 
+    with pytest.raises(
+        ValueError,
+        match=r'^Each reference scene must have a unique burst number \+ swath \+ polarization identifier$',
+    ):
+        burst.validate_bursts(
+            [
+                'S1_000000_IW1_20200101T000000_VV_0000-BURST',
+                'S1_000000_IW1_20200101T000001_VV_0000-BURST',
+            ],
+            [
+                'S1_000000_IW1_20200201T000000_VV_0000-BURST',
+                'S1_000000_IW1_20200201T000002_VV_0000-BURST',
+            ],
+        )
+
     with pytest.raises(ValueError, match=r'^Scenes must have the same polarization. Polarizations present: VH, VV$'):
         burst.validate_bursts(
             [
@@ -153,11 +168,11 @@ def test_validate_bursts_num_swath_pol():
         burst.validate_bursts(
             [
                 'S1_000000_IW1_20200101T000000_VH_0000-BURST',
-                'S1_000000_IW1_20200101T000000_VH_0000-BURST',
+                'S1_000000_IW2_20200101T000000_VH_0000-BURST',
             ],
             [
                 'S1_000000_IW1_20200201T000000_VH_0000-BURST',
-                'S1_000000_IW1_20200201T000000_VH_0000-BURST',
+                'S1_000000_IW2_20200201T000000_VH_0000-BURST',
             ],
         )