+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - Reference and secondary scenes are now accepted in any order, but must still have the same set of burst number + swath + polarization identifiers.
-- Pinned to `burst2safe>=1.4.5` to address https://github.com/ASFHyP3/burst2safe/issues/170.
 
 ### Fixed
+- Reference scenes are now rejected if they contain any duplicate burst number + swath + polarization identifiers.
 - Allow scenes over water when water masking is applied. Fixes https://github.com/ASFHyP3/hyp3-isce2/issues/211 and https://github.com/ASFHyP3/hyp3-isce2/issues/247.
+- Pinned to `burst2safe>=1.4.5` to address https://github.com/ASFHyP3/burst2safe/issues/170.
 
 ## [2.3.1]